First version of the new API with Promises. Breaks the old API
This is the first of many steps towards Node-Mpv Version 2. Calling the constructor only initializes the MPV object now, but does not start the MPV player yet.
One has to call the start method to do this. The start method returns a promise which resolves if the IPC Socket could succesfully be bound and is rejected if not.