Fix for IPCRequest ID when using latest version (git) of mpv
In mpv's latest iterations (at least as of this commit's date, perhaps
it happens before), request_id is always present in requests when data
is null and contains 0.

Before the property just did not exist.

It caused an uncaught exception as it could not find the IPC Request ID
in the IPCRequest instance.