play() can now start playback from idle state. Closes #64
When mpv is started with --idle (which Node-MPV does), and files are added using 'append', the only way to start the playback later is by setting 'playlist-pos' to 0. Since this is very unuintuitive, I implemented this into the play() function. It checks if mpv is idle and the playlist size is bigger than 0, if so it sets 'playlist-pos' to 0 to start playback, if not it just sets the pause state to false, as usualy